requirements.txt
joblib==1.2.0
matplotlib==3.8.0
numpy
pandas==2.2.2
plotly==5.9.0
scikit_learn==1.4.2
scipy==1.13.1
streamlit==1.35.0
streamlit_elements==0.1.0
streamlit_toggle_switch==1.0.2
xgboost==2.0.3
openpyxl==3.1.2
xlrd==2.0.1
streamlit-aggrid
xgboost

